refactor date picker files to be more flexiable and readable
