+++ app/src/main/java/com/baghdad/tudee/MainActivity.kt
@@ -4,14 +4,16 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import com.baghdad.tudee.designSystem.theme.TudeeTheme
+import com.baghdad.tudee.ui.composable.dateYearDialog.DateDialog
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -19,15 +21,19 @@ class MainActivity : ComponentActivity() {
         enableEdgeToEdge()
         setContent {
             TudeeTheme {
-                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    Greeting(
-                        name = "Android",
-                        modifier = Modifier.padding(innerPadding)
-                    )
-                }
+                var showDatePicker by remember { mutableStateOf(false) }
+                var selectedDateMillis by remember { mutableStateOf<Long?>(null) }
+                DateDialog(
+                    isShowDatePicker = showDatePicker,
+                    onDismiss = { showDatePicker = false },
+                    onDateSelected = { millis ->
+                        selectedDateMillis = millis
+
+                    })
             }
         }
     }
+
 }
 
 @Composable
